refactor: update secureNoteType filename
