+++ .github/whitelist-capital-letters.txt
@@ -39,7 +39,6 @@
 ./libs/common/src/misc/serviceUtils.ts
 ./libs/common/src/misc/serviceUtils.spec.ts
 ./libs/common/src/enums/uriMatchType.ts
-./libs/common/src/enums/secureNoteType.ts
 ./libs/common/src/enums/storageLocation.ts
 ./libs/common/src/enums/themeType.ts
 ./libs/common/src/enums/stateVersion.ts
