+++ libs/common/src/enums/secure-note-type.ts
