+++ libs/common/src/models/export/secure-note.export.ts
@@ -1,4 +1,4 @@
-import { SecureNoteType } from "../../enums/secureNoteType";
+import { SecureNoteType } from "../../enums/secure-note-type";
 import { SecureNote as SecureNoteDomain } from "../../vault/models/domain/secure-note";
 import { SecureNoteView } from "../../vault/models/view/secure-note.view";
