+++ libs/common/src/vault/models/data/secure-note.data.ts
@@ -1,4 +1,4 @@
-import { SecureNoteType } from "../../../enums/secureNoteType";
+import { SecureNoteType } from "../../../enums/secure-note-type";
 import { SecureNoteApi } from "../../../models/api/secure-note.api";
 
 export class SecureNoteData {
