+++ libs/common/src/vault/models/view/secure-note.view.ts
@@ -1,6 +1,6 @@
 import { Jsonify } from "type-fest";
 
-import { SecureNoteType } from "../../../enums/secureNoteType";
+import { SecureNoteType } from "../../../enums/secure-note-type";
 import { SecureNote } from "../domain/secure-note";
 
 import { ItemView } from "./item.view";
