+++ libs/importer/src/importers/bitwarden/bitwarden-csv-importer.ts
@@ -1,6 +1,6 @@
 import { CollectionView } from "@bitwarden/common/admin-console/models/view/collection.view";
 import { FieldType } from "@bitwarden/common/enums/field-type";
-import { SecureNoteType } from "@bitwarden/common/enums/secureNoteType";
+import { SecureNoteType } from "@bitwarden/common/enums/secure-note-type";
 import { CipherRepromptType } from "@bitwarden/common/vault/enums/cipher-reprompt-type";
 import { CipherType } from "@bitwarden/common/vault/enums/cipher-type";
 import { CipherView } from "@bitwarden/common/vault/models/view/cipher.view";
