+++ libs/importer/src/importers/onepassword/onepassword-1pif-importer.ts
@@ -1,5 +1,5 @@
 import { FieldType } from "@bitwarden/common/enums/field-type";
-import { SecureNoteType } from "@bitwarden/common/enums/secureNoteType";
+import { SecureNoteType } from "@bitwarden/common/enums/secure-note-type";
 import { CipherType } from "@bitwarden/common/vault/enums/cipher-type";
 import { CardView } from "@bitwarden/common/vault/models/view/card.view";
 import { CipherView } from "@bitwarden/common/vault/models/view/cipher.view";
